requirements.txt
fillpdf
aiofiles==24.1.0
aioredis==1.3.1
aiosmtplib==3.0.2
aiosqlite==0.17.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
apache-libcloud==3.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
async-timeout==4.0.3
asyncpg==0.29.0
Authlib==1.3.2
babel==2.16.0
bcrypt==4.1.2
boto3==1.35.49
botocore==1.35.99
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
crypto==1.4.1
cryptography==43.0.0
cssselect2==0.8.0
databases==0.9.0
dnspython==2.6.1
docxcompose==1.4.0
docxtpl==0.18.0
email_validator==2.1.2
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-admin==1.0.4
fastapi-cli==0.0.4
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
fillpdf==0.7.3
fonttools==4.58.0
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iso8601==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
lxml==5.3.0
makefun==1.15.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
Naked==0.1.32
orjson==3.10.6
packaging==24.1
passlib==1.7.4
pdf2image==1.17.0
pdfrw2==0.5.0
pendulum==3.0.0
pillow==11.2.1
pwdlib==0.2.0
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.3.4
pydantic_core==2.20.1
pydyf==0.11.0
Pygments==2.18.0
PyJWT==2.8.0
PyMuPDF==1.25.5
pyphen==0.17.2
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
s3transfer==0.10.4
shellescape==3.8.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqladmin==0.18.0
SQLAlchemy==2.0.32
sqlalchemy-file==0.6.0
starlette==0.37.2
time-machine==2.15.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tortoise-orm==0.21.5
tqdm==4.66.6
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
weasyprint==65.1
webencodings==0.5.1
websockets==12.0
WTForms==3.1.2
zopfli==0.2.3.post1
